Consolidate common code to IO mod
