+++ common/src/lib.rs
@@ -31,3 +31,4 @@ pub mod error;
 pub mod config;
 pub mod flakelog;
 pub mod defaults;
+pub mod io;
