+++ firecracker-pilot/src/firecracker.rs
@@ -23,6 +23,7 @@
 //
 use std::ffi::OsStr;
 use std::{thread, time};
+use flakes::io::IO;
 use flakes::command::{CommandError, handle_output, CommandExtTrait};
 use flakes::error::{FlakeError, OperationError};
 use flakes::user::{User, mkdir};
@@ -150,9 +151,15 @@ pub fn create(program_name: &String) -> Result<(String, String), FlakeError> {
           # Optional path to initrd image done by app registration
           initrd_path: /var/lib/firecracker/images/NAME/initrd
 
-      Calling this method returns a vector including a placeholder
-      for the later VM process ID and and the name of
-      the VM ID file.
+      include:
+        tar:
+          - tar-archive-file-name-to-include
+        path:
+          - file-or-directory-to-include
+
+    Calling this method returns a vector including a placeholder
+    for the later VM process ID and and the name of
+    the VM ID file.
     !*/
     if ! Lookup::which(defaults::FIRECRACKER) {
         return Err(FlakeError::IOError {
@@ -274,7 +281,9 @@ fn run_creation(
                 if Lookup::is_debug() {
                     debug!("Syncing includes...");
                 }
-                sync_includes(&vm_mount_point, User::ROOT)?;
+                IO::sync_includes(
+                    &vm_mount_point, config().tars(), config().paths(), User::ROOT
+                )?;
             }
             umount_vm(tmp_dir, User::ROOT)?;
         }
@@ -820,69 +829,6 @@ pub fn delete_file(filename: &String, user: User) -> bool {
         }
     }
     true
-}
-
-pub fn sync_includes(
-    target: &String, user: User
-) -> Result<(), FlakeError> {
-    /*!
-    Sync custom include data to target path
-    !*/
-    let tar_includes = &config().tars();
-    let path_includes = &config().paths();
-
-    for tar in tar_includes {
-        if Lookup::is_debug() {
-            debug!("Provision tar archive: [{}]", tar);
-        }
-        let mut call = user.run("tar");
-        call.arg("-C").arg(target)
-            .arg("-xf").arg(tar);
-        if Lookup::is_debug() {
-            debug!("{:?}", call.get_args());
-        }
-        let output = call.perform()?;
-        if Lookup::is_debug() {
-            debug!("{}", &String::from_utf8_lossy(&output.stdout));
-            debug!("{}", &String::from_utf8_lossy(&output.stderr));
-        }
-    }
-    for path in path_includes {
-        if Lookup::is_debug() {
-            debug!("Provision path: [{}]", path);
-        }
-        sync_data(
-            path, &format!("{}/{}", target, path),
-            ["--mkpath"].to_vec(), user
-        )?;
-    }
-    Ok(())
-}
-
-pub fn sync_data(
-    source: &str, target: &str, options: Vec<&str>, user: User
-) -> Result<(), FlakeError> {
-    /*!
-    Sync data from source path to target path
-    !*/
-    let mut call = user.run("rsync");
-    call.arg("-av");
-    for option in options {
-        call.arg(option);
-    }
-    call.arg(source).arg(target);
-    if Lookup::is_debug() {
-        debug!("{:?}", call.get_args());
-    }
-    let output = call.output()?;
-    if Lookup::is_debug() {
-        debug!("{}", &String::from_utf8_lossy(&output.stdout));
-        debug!("{}", &String::from_utf8_lossy(&output.stderr));
-    }
-    if !output.status.success() {
-        return Err(FlakeError::SyncFailed)
-    }
-    Ok(())
 }
 
 pub fn mount_vm(
